requirements.txt
django>=4.0
djangorestframework>=3.13
psycopg2-binary>=2.9
python-dotenv>=0.19.0
djangorestframework>=3.13.0
